Update AutoresController.php, EditorasController.php, and 4 more files...

+++ app/Http/Controllers/AutoresController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use App\Models\Autor;
 
 class AutoresController extends Controller
@@ -27,9 +28,15 @@ class AutoresController extends Controller
     	]);
     }
     public function create() {
-        return view('autores.create');
+        if (Gate::allows('admin')) {
+            return view('autores.create');
+        }
+        else{
+            return redirect()->route('autores.index')->with('mensagem', 'Erro não tem permissoes para entrar nesta area');
+        }
     }
     public function store(Request $request){
+        if (Gate::allows('admin')) {
         $novoAutor = $request->validate([
             'nome'=>['required', 'min:1', 'max:50'],
             'nacionalidade'=> ['nullable','min:3', 'max:50'],
@@ -39,17 +46,28 @@ class AutoresController extends Controller
         $autor =Autor::create($novoAutor);
         return redirect()->route('autores.show', [
                 'id'=>$autor->id_autor
-            ]);
+            ]);            
+        }
+        else{
+            return redirect()->route('autores.index')->with('mensagem', 'Erro não tem permissoes para entrar nesta area');
+        }
+
     }
     public function edit(Request $request) {
+        if(Gate::allows('admin')) {
         $idAutor=$request->id;
         $idAutor = Autor::where('id_autor', $idAutor)->first();
         return view('autores.edit', [
             'autor'=>$idAutor
-        ]);
+        ]); 
+        }
+        else{
+            return redirect()->route('autores.index')->with('mensagem', 'Erro não tem permissoes para entrar nesta area');
+        }
     }
     public function update (Request $request) {
-        $idAutor=$request->id;
+        if(Gate::allows('admin')){
+           $idAutor=$request->id;
         $autor = Autor::findOrFail($idAutor);
             $atualizarAutor = $request->validate ([
             'nome'=>['required', 'min:1', 'max:50'],
@@ -60,10 +78,15 @@ class AutoresController extends Controller
         $autor->update($atualizarAutor);   
         return redirect()->route('autores.show', [
             'id'=>$autor->id_autor
-        ]); 
+        ]);  
+        }
+        else{
+            return redirect()->route('autores.index')->with('mensagem', 'Erro não tem permissoes para entrar nesta area');
+        }
     }
     public function delete(Request $request) {
-        $idAutor=$request->id;
+        if(Gate::allows('admin')) {
+            $idAutor=$request->id;
         $autor=Autor::where('id_autor', $idAutor )->first();
         if(is_null($autor)) {
             return redirect()->route('autores.index')
@@ -72,13 +95,23 @@ class AutoresController extends Controller
         else {
             return view('autores.delete', ['autor'=>$autor]);
         }
+        }
+        else{
+            return redirect()->route('autores.index')->with('mensagem', 'Erro não tem permissoes para entrar nesta area');
+        }
+        
     }
     public function destroy(Request $request) {
-        $idAutor=$request->id;
-        $autor =Autor::findOrFail($idAutor);
-        $autor->delete();
-        return redirect()->route('autores.index')
-        ->with('mensagem', 'Autor Eliminado');
+        if(Gate::allows('admin')) {
+            $idAutor=$request->id;
+            $autor =Autor::findOrFail($idAutor);
+            $autor->delete();
+            return redirect()->route('autores.index')
+            ->with('mensagem', 'Autor Eliminado');
+        }
+        else {
+            return redirect()->route('autores.index')->with('mensagem', 'Erro não tem permissoes para entrar nesta area');
+        }     
     }
 
 }
