+++ app/Http/Controllers/GenerosController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use App\Models\Genero;
 
 class GenerosController extends Controller
@@ -27,10 +28,16 @@ class GenerosController extends Controller
     	]);
     }
     public function create() {
-        return view('generos.create');
+        if (Gate::allows('admin')) {
+            return view('generos.create');
+        }
+        else{
+            return redirect()->route('generos.index')->with('mensagem', 'Erro não tem permissoes para entrar nesta area');
+        }
     }
     public function store(Request $request){
-        $novoGenero = $request->validate([
+        if (Gate::allows('admin')) {
+            $novoGenero = $request->validate([
             'designacao'=>['required', 'min:1', 'max:50'],
             'observacoes'=> ['nullable','min:3', 'max:200'],
         ]);
@@ -38,16 +45,27 @@ class GenerosController extends Controller
         return redirect()->route('generos.show', [
             'id'=>$genero->id_genero
         ]);
+        }
+        else{
+            return redirect()->route('generos.index')->with('mensagem', 'Erro não tem permissoes para entrar nesta area');
+        }
+        
     }
     public function edit(Request $request) {
-        $idGenero=$request->id;
+        if(Gate::allows('admin')) {
+            $idGenero=$request->id;
         $idGenero = Genero::where('id_genero', $idGenero)->first();
         return view('generos.edit', [
             'genero'=>$idGenero
         ]);
+        }
+        else{
+            return redirect()->route('generos.index')->with('mensagem', 'Erro não tem permissoes para entrar nesta area');
+        }  
     }
     public function update (Request $request) {
-        $idGenero=$request->id;
+        if (Gate::allows('admin')) {
+            $idGenero=$request->id;
         $genero = Genero::findOrFail($idGenero);
             $atualizarGenero = $request->validate ([
             'designacao'=>['required', 'min:1', 'max:50'],
@@ -56,10 +74,16 @@ class GenerosController extends Controller
         $genero->update($atualizarGenero);   
         return redirect()->route('generos.show', [
             'id'=>$genero->id_genero
-        ]); 
+        ]);
+        }
+        else{
+            return redirect()->route('generos.index')->with('mensagem', 'Erro não tem permissoes para entrar nesta area');
+        }
+         
     }
     public function delete(Request $request) {
-        $idGenero=$request->id;
+        if (Gate::allows('admin')) {
+           $idGenero=$request->id;
         $genero=Genero::where('id_genero', $idGenero )->first();
         if(is_null($genero)) {
             return redirect()->route('generos.index')
@@ -67,13 +91,24 @@ class GenerosController extends Controller
         }
         else {
             return view('generos.delete', ['genero'=>$genero]);
+        } 
         }
+        else{
+            return redirect()->route('generos.index')->with('mensagem', 'Erro não tem permissoes para entrar nesta area');
+        }
+        
     }
     public function destroy(Request $request) {
-        $idGenero=$request->id;
-        $genero =Genero::findOrFail($idGenero);
-        $genero->delete();
-        return redirect()->route('generos.index')
-        ->with('mensagem', 'Genero Eliminado');
+        if(Gate::allows('admin')) {
+            $idGenero=$request->id;
+            $genero =Genero::findOrFail($idGenero);
+            $genero->delete();
+            return redirect()->route('generos.index')
+            ->with('mensagem', 'Genero Eliminado');
+        }
+        else{
+            return redirect()->route('generos.index')->with('mensagem', 'Erro não tem permissoes para entrar nesta area');
+        }
+        
     }
 }
